+++ src/mulle-bashfunctions-all.sh
@@ -2138,6 +2138,53 @@ function r_escaped_shell_string()
 }
 
 
+function r_escaped_json()
+{
+    RVAL="${*//\\/\\\\}"
+
+    RVAL="${RVAL//\"/\\\"}"
+
+    RVAL="${RVAL//$'\n'/\\n}"
+
+    RVAL="${RVAL//$'\r'/\\r}"
+
+    RVAL="${RVAL//$'\t'/\\t}"
+
+    RVAL="${RVAL//$'\b'/\\b}"
+
+    RVAL="${RVAL//$'\f'/\\f}"
+
+    RVAL="${RVAL//\//\\/}"
+
+    RVAL="${RVAL//$'\x1'/\\u0001}"
+    RVAL="${RVAL//$'\x2'/\\u0002}"
+    RVAL="${RVAL//$'\x3'/\\u0003}"
+    RVAL="${RVAL//$'\x4'/\\u0004}"
+    RVAL="${RVAL//$'\x5'/\\u0005}"
+    RVAL="${RVAL//$'\x6'/\\u0006}"
+    RVAL="${RVAL//$'\x7'/\\u0007}"
+    RVAL="${RVAL//$'\xB'/\\u000B}"
+    RVAL="${RVAL//$'\xE'/\\u000E}"
+    RVAL="${RVAL//$'\xF'/\\u000F}"
+    RVAL="${RVAL//$'\x10'/\\u0010}"
+    RVAL="${RVAL//$'\x11'/\\u0011}"
+    RVAL="${RVAL//$'\x12'/\\u0012}"
+    RVAL="${RVAL//$'\x13'/\\u0013}"
+    RVAL="${RVAL//$'\x14'/\\u0014}"
+    RVAL="${RVAL//$'\x15'/\\u0015}"
+    RVAL="${RVAL//$'\x16'/\\u0016}"
+    RVAL="${RVAL//$'\x17'/\\u0017}"
+    RVAL="${RVAL//$'\x18'/\\u0018}"
+    RVAL="${RVAL//$'\x19'/\\u0019}"
+    RVAL="${RVAL//$'\x1A'/\\u001A}"
+    RVAL="${RVAL//$'\x1B'/\\u001B}"
+    RVAL="${RVAL//$'\x1C'/\\u001C}"
+    RVAL="${RVAL//$'\x1D'/\\u001D}"
+    RVAL="${RVAL//$'\x1E'/\\u001E}"
+    RVAL="${RVAL//$'\x1F'/\\u001F}"
+ }
+
+
 
 function string_has_prefix()
 {
@@ -3903,6 +3950,7 @@ _r_make_tmp_in_dir()
    name="${name:-${MULLE_EXECUTABLE_NAME}}"
    name="${name:-mulle}"
 
+   extension="${extension#\.}"
    if [ ! -z "${extension}" ]
    then
       extension=".${extension}"